move `build-worker` functionality to `terraform-worker` remove `build-worker` rename `terraform-worker` --> `worker`
